sql/util/requirements.txt
pandas==2.3.3
google-cloud-bigquery==3.40.0
google-api-python-client==2.188.0
google-auth==2.47.0
requests==2.32.5
tabulate==0.9.0
gspread==6.2.1
gspread-dataframe==4.0.0
ipykernel==7.1.0
db-dtypes==1.5.0
